refactor: move language service into separate module for better separation for concerns
