+++ packages/extension/src/a.js
@@ -1,0 +1,3 @@
+import * as a from 'language-service'
+
+console.log(a)
