+++ packages/extension/src/languageFeaturesCssMain.js
@@ -1,7 +1,7 @@
 // TODO would be nice to have typescript for all extensions
 // but is it worth the compile time?
-import * as ExtensionHostCompletionProviderCss from './parts/ExtensionHost/ExtensionHostCompletionProviderCss.js'
-import * as ExtensionHostTabCompletionProviderCss from './parts/ExtensionHost/ExtensionHostTabCompletionProviderCss.js'
+import * as ExtensionHostCompletionProviderCss from './parts/ExtensionHostCompletion/ExtensionHostCompletion.js'
+import * as ExtensionHostTabCompletionProviderCss from './parts/ExtensionHostTabCompletion/ExtensionHostTabCompletion.js'
 
 export const activate = () => {
   // TODO could also define completionProvider = { languageId: 'html', providerCompletions(){} }
