+++ packages/language-service-data/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "language-service-data",
+  "version": "0.0.0-dev",
+  "description": "",
+  "main": "index.js",
+  "scripts": {},
+  "keywords": [],
+  "author": "",
+  "license": "MIT"
+}
