+++ packages/language-service-data/src/css-properties.json
