+++ packages/language-service-data/src/propertyData.js
