+++ packages/language-service-data/src/tabCompletionSnippets.js
