+++ packages/language-service/package.json
@@ -1,17 +1,18 @@
 {
-  "name": "builtin.language-features-css",
+  "name": "language-service",
   "version": "0.0.0-dev",
   "description": "",
-  "main": "src/languageFeaturesCssMain.js",
+  "main": "index.js",
+  "type": "module",
   "scripts": {
     "test": "node --unhandled-rejections=warn --experimental-vm-modules ./node_modules/jest/bin/jest.js --detectOpenHandles --forceExit",
     "test:watch": "node --unhandled-rejections=warn --experimental-vm-modules ./node_modules/jest/bin/jest.js --watch"
   },
   "keywords": [],
   "author": "",
-  "type": "module",
+  "license": "MIT",
   "dependencies": {
-    "language-service": "file:../language-service"
+    "language-service-data": "file:../language-service-data"
   },
   "devDependencies": {
     "@types/jest": "^29.0.0",
