+++ packages/language-service/src/languageServiceMain.js
@@ -1,0 +1,2 @@
+export * as Completion from './parts/Completion/Completion.js'
+export * as TabCompletion from './parts/TabCompletion/TabCompletion.js'
