+++ packages/language-service/src/parts/GetMatchingCompletion/GetMatchingCompletion.js
@@ -1,10 +1,18 @@
 import { readFileSync } from 'node:fs'
 import { join } from 'node:path'
-import * as Root from './Root/Root.js'
+import * as Root from '../Root/Root.js'
 
 // TODO use import json once supported
 const cssProperties = JSON.parse(
-  readFileSync(join(Root.root, 'data/css-properties.json')).toString()
+  readFileSync(
+    join(
+      Root.root,
+      'packages',
+      'language-service-data',
+      'src',
+      'css-properties.json'
+    )
+  ).toString()
 )
 
 const merge = (objectA, objectB) => ({ ...objectA, ...objectB })
