+++ packages/language-service/src/parts/Root/Root.js
@@ -1,8 +1,7 @@
-import { readFileSync } from 'node:fs'
 import { dirname, join } from 'node:path'
 import { fileURLToPath } from 'node:url'
 
 // TODO use import json once supported
 const __dirname = dirname(fileURLToPath(import.meta.url))
 
-export const root = join(__dirname, '..', '..', '..')
+export const root = join(__dirname, '..', '..', '..', '..', '..')
