+++ packages/language-service/test/Completion.test.js
