+++ packages/language-service/tsconfig.json
@@ -13,5 +13,5 @@
     "assumeChangesOnlyAffectDirectDependencies": true,
     "strict": true
   },
-  "include": ["src", "test", "../language-service-data/data"]
+  "include": ["src", "test"]
 }
